moved constants into migrator instances

+++ nislmigrate/facades/file_system_facade.py
@@ -4,6 +4,7 @@ import os
 import shutil
 import stat
 
+from nislmigrate.logging.migration_error import MigrationError
 from nislmigrate.migration_action import MigrationAction
 
 
@@ -82,7 +83,6 @@ class FileSystemFacade:
         if os.path.isdir(dir_):
             shutil.rmtree(dir_, onerror=self.remove_readonly)
 
-# TODO: make a migrate directory again.
     def migrate_singlefile(self,
                            migration_directory_root: str,
                            service_name: str,
@@ -135,14 +135,34 @@ class FileSystemFacade:
         singlefile_full_path = os.path.join(migration_dir, file)
         shutil.copy(singlefile_full_path, restore_directory)
 
-    def copy_directory(self,
-                       from_directory: str,
-                       to_directory: str):
+    @staticmethod
+    def copy_file(from_directory: str, to_directory: str, file_name: str) -> None:
         """
         Copy an entire directory from one location to another.
 
-        :param from_directory: The name of the service.
-        :param to_directory: Whether to capture or restore.
+        :param from_directory: The directory the file to copy exists in.
+        :param to_directory: The directory to copy the file into.
+        :param file_name: The name of the file to copy.
         """
+        # TODO: Might not need this if copy creates a directory.
+        if not os.path.exists(to_directory):
+            os.mkdir(to_directory)
+        file_path = os.path.join(from_directory, file_name)
+        shutil.copy(file_path, to_directory)
+
+    def copy_directory(self, from_directory: str, to_directory: str, force: bool):
+        """
+        Copy an entire directory from one location to another.
+
+        :param from_directory: The directory whose contents to copy.
+        :param to_directory: The directory to put the copied contents.
+        :param force: Whether to delete existing content in to_directory before copying.
+        """
+        if os.path.exists(to_directory) and os.listdir(to_directory) and not force:
+            error = "The tool can not copy to the non empty directory: '%s'" % to_directory
+            raise MigrationError(error)
+        if not os.path.exists(from_directory):
+            raise MigrationError("No data found at: '%s'" % from_directory)
+
         self.remove_dir(to_directory)
-        shutil.copytree(from_directory, to_directory)
+        shutil.copytree(from_directory, to_directory, dirs_exist_ok=False)
